Initial commit
Tests seem to work both in vscode as well as via cmake
Generation of the gRPC and protobuf files works via cmake also
All the code is in python now.  Just a couple of services ATM:

terrain_generation:  makes a block of land in the form of hexagons
	with varying environmental types
persistence:  python/SQLAlchemy over sqlite for now.