11-introduce-an-api-gateway-to-simplify-ui-development
Envoy is the gateway and we are using it via its container published on
docker hub.

The gateway does GRPC proxying rather than providing REST to clients and
doing GRPC to the backend services.  The hope is that this will provide
better run time performance for clients.

Start and stop scripts were updated to handle the envoy container
properly.

GRPC reflection APIs were enabled on enoyv and the tewo backend services
to make debugging with grpccurl possible.