3-create-a-ui-in-react
The creation of the react app made clear that the gateway config for
envoy wasn't going to work with browsers (mainly due to CORS).
Lots of misery getting there, but now that's fixed.  The react app is
roughly the equivalent of the earlier python vie_ui.py.  Along the way,
the dream of supporting GRPC clients (like the python client) and
GRPC-web clients (like the new react one) died a painful death.  As
it turns out, GRPC for web clients is HTTP2 only and support for that
is spotty.  So this change also transitions us to GRPC-web as the
only supported set of interfaces for clients.