164-consolidate-logs-to-stdout-and-then-send-that-to-a-log-file
Done.  Also massaged the log formats so envoy's would look of a kind
with the rest.  Currently the logs all go to stdout and then are
redirected to a time-stamped-named file in ./logs.  That last bit
happens in start.sh, so perfect for development use cases while
still allowing for the raw stdout to be sent wherever it needs to go
in cloud deployments.